backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
requests>=2.31.0
python-dotenv==1.0.0
pydub>=0.25.1
mutagen>=1.47.0
librosa>=0.10.0
openai>=1.0.0
# Legacy dependencies (kept for backward compatibility)
# faster-whisper==1.0.3
# pyannote.audio==3.1.1
# torch>=1.9.0
# torchaudio>=0.9.0
# assemblyai>=1.0.0  # No longer needed, using direct HTTP API
